Adding sub directories to hook directories (#183)
