+++ src/components/CustomCall.tsx
@@ -20,7 +20,7 @@ import { Message } from 'semantic-ui-react';
 import { ButtonSubmit } from '../components';
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useTransactionContext } from '../contexts/TransactionContext';
-import useSendMessage from '../hooks/useSendMessage';
+import useSendMessage from '../hooks/chain/useSendMessage';
 import { TransactionTypes } from '../types/transactionTypes';
 
 const CustomCall = () => {
