+++ src/components/Remark.tsx
@@ -19,7 +19,7 @@ import React, { useState } from 'react';
 import { ButtonSubmit } from '../components';
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useTransactionContext } from '../contexts/TransactionContext';
-import useSendMessage from '../hooks/useSendMessage';
+import useSendMessage from '../hooks/chain/useSendMessage';
 import { TransactionTypes } from '../types/transactionTypes';
 
 const Remark = () => {
