+++ src/components/Transfer.tsx
@@ -19,9 +19,9 @@ import { Box, makeStyles, TextField, Typography } from '@material-ui/core';
 import BN from 'bn.js';
 import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useTransactionContext } from '../contexts/TransactionContext';
-import useAccounts from '../hooks/useAccounts';
-import useBalance from '../hooks/useBalance';
-import useSendMessage from '../hooks/useSendMessage';
+import useAccounts from '../hooks/accounts/useAccounts';
+import useBalance from '../hooks/subscriptions/useBalance';
+import useSendMessage from '../hooks/chain/useSendMessage';
 import { TransactionTypes } from '../types/transactionTypes';
 import { TokenSymbol } from './TokenSymbol';
 import Receiver from './Receiver';
