+++ src/hooks/accounts/useAccounts.ts
@@ -14,19 +14,19 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity Bridges UI.  If not, see <http://www.gnu.org/licenses/>.
 
-import { TransactionActionCreators } from '../actions/transactionActions';
+import { TransactionActionCreators } from '../../actions/transactionActions';
 import type { KeyringPair } from '@polkadot/keyring/types';
 import { encodeAddress } from '@polkadot/util-crypto';
 import { useEffect, useState } from 'react';
-import { useUpdateTransactionContext } from '../contexts/TransactionContext';
-import { AccountActionCreators } from '../actions/accountActions';
-import { SourceTargetActionsCreators } from '../actions/sourceTargetActions';
-import { useUpdateAccountContext } from '../contexts/AccountContextProvider';
-import { useAccountContext } from '../contexts/AccountContextProvider';
-import { useKeyringContext } from '../contexts/KeyringContextProvider';
-import { useUpdateSourceTarget } from '../contexts/SourceTargetContextProvider';
-import useDerivedAccount from '../hooks/useDerivedAccount';
-import useChainGetters from '../hooks/useChainGetters';
+import { useUpdateTransactionContext } from '../../contexts/TransactionContext';
+import { AccountActionCreators } from '../../actions/accountActions';
+import { SourceTargetActionsCreators } from '../../actions/sourceTargetActions';
+import { useUpdateAccountContext } from '../../contexts/AccountContextProvider';
+import { useAccountContext } from '../../contexts/AccountContextProvider';
+import { useKeyringContext } from '../../contexts/KeyringContextProvider';
+import { useUpdateSourceTarget } from '../../contexts/SourceTargetContextProvider';
+import useDerivedAccount from './useDerivedAccount';
+import useChainGetters from '../chain/useChainGetters';
 
 interface Accounts {
   account: KeyringPair | null;
