+++ src/hooks/accounts/useDerivedAccount.ts
@@ -13,27 +13,31 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with Parity Bridges UI.  If not, see <http://www.gnu.org/licenses/>.
-import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 
-interface LoadingStates {
-  areApiReady: boolean;
-  sourceReady: boolean;
-  targetReady: boolean;
-}
+import { useAccountContext } from '../../contexts/AccountContextProvider';
+import { useSourceTarget } from '../../contexts/SourceTargetContextProvider';
+import { getBridgeId } from '../../util/getConfigs';
+import getDeriveAccount from '../../util/getDeriveAccount';
 
-export default function useLoadingApi(): LoadingStates {
+const useDerivedAccount = () => {
   const {
+    targetChainDetails: { configs },
     sourceChainDetails: {
-      apiConnection: { isApiReady: isSourceApiReady }
-    },
-    targetChainDetails: {
-      apiConnection: { isApiReady: isTargetApiReady }
+      configs: { chainName }
     }
   } = useSourceTarget();
+  const { account } = useAccountContext();
 
-  return {
-    areApiReady: isSourceApiReady && isTargetApiReady,
-    sourceReady: isSourceApiReady,
-    targetReady: isTargetApiReady
+  if (!account) {
+    return null;
+  }
+
+  const toDerive = {
+    ss58Format: configs.ss58Format,
+    address: account.address,
+    bridgeId: getBridgeId(configs, chainName)
   };
-}
+  return getDeriveAccount(toDerive);
+};
+
+export default useDerivedAccount;
