+++ src/hooks/chain/useSendMessage.ts
@@ -19,18 +19,18 @@ import { web3FromSource } from '@polkadot/extension-dapp';
 import type { KeyringPair } from '@polkadot/keyring/types';
 import moment from 'moment';
 
-import { MessageActionsCreators } from '../actions/messageActions';
-import { TransactionActionCreators } from '../actions/transactionActions';
-import { useAccountContext } from '../contexts/AccountContextProvider';
-import { useUpdateMessageContext } from '../contexts/MessageContext';
-import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
-import { useTransactionContext, useUpdateTransactionContext } from '../contexts/TransactionContext';
-import useLaneId from '../hooks/useLaneId';
-import useTransactionPreparation from '../hooks/useTransactionPreparation';
-import { TransactionStatusEnum, TransactionTypes } from '../types/transactionTypes';
-import getSubstrateDynamicNames from '../util/getSubstrateDynamicNames';
-import logger from '../util/logger';
-import useLoadingApi from '../hooks/useLoadingApi';
+import { MessageActionsCreators } from '../../actions/messageActions';
+import { TransactionActionCreators } from '../../actions/transactionActions';
+import { useAccountContext } from '../../contexts/AccountContextProvider';
+import { useUpdateMessageContext } from '../../contexts/MessageContext';
+import { useSourceTarget } from '../../contexts/SourceTargetContextProvider';
+import { useTransactionContext, useUpdateTransactionContext } from '../../contexts/TransactionContext';
+import useLaneId from './useLaneId';
+import useTransactionPreparation from '../transactions/useTransactionPreparation';
+import { TransactionStatusEnum, TransactionTypes } from '../../types/transactionTypes';
+import getSubstrateDynamicNames from '../../util/getSubstrateDynamicNames';
+import logger from '../../util/logger';
+import useLoadingApi from '../connections/useLoadingApi';
 
 interface Props {
   isValidCall?: boolean;
