+++ src/hooks/connections/useApiConnection.ts
@@ -18,9 +18,9 @@ import { ApiPromise } from '@polkadot/api';
 import { TypeRegistry } from '@polkadot/types';
 import isEmpty from 'lodash/isEmpty';
 import React, { useEffect, useState } from 'react';
-import { ApiPromiseConnectionType, Configs, ConnectionChainInformation } from '../types/sourceTargetTypes';
-import logger from '../util/logger';
-import { getConfigs } from '../util/getConfigs';
+import { ApiPromiseConnectionType, Configs, ConnectionChainInformation } from '../../types/sourceTargetTypes';
+import logger from '../../util/logger';
+import { getConfigs } from '../../util/getConfigs';
 
 export const ApiPromiseContext: React.Context<ApiPromiseConnectionType> = React.createContext(
   {} as ApiPromiseConnectionType
