+++ src/hooks/subscriptions/useApiSubscription.ts
@@ -16,7 +16,7 @@
 
 import { UnsubscribePromise } from '@polkadot/api/types';
 import { useEffect } from 'react';
-import logger from '../util/logger';
+import logger from '../../util/logger';
 
 export function useApiSubscription(fn: () => UnsubscribePromise, isReady: boolean): void {
   useEffect(() => {
